Add files via upload
 LANGUAGES USED

•	PYTHON
•	MACHINE LEARNING METHODS.


 TECHNICAL REQUIREMENTS(HARDWARE/SOFTWARE)

•	JETSON NANO
•	JUPYTER NOTEBOOK
•	PYTHON IDE

RESULTS ACHIEVED

                                                                  Table 2: PCA DATA SET

             METHODS	  TRAIN ACCURACY	      TEST ACCURACY
                   SVM	              0.80	                 0.84 
                    RF	              0.80	                 0.95
                    LR	              0.80	                 0.92

 CONCLUSION

In this work, we mainly solve the problem of fast classification of lidar signals for measuring air pollution.
By comparison, it can be concluded that the classification accuracy can be improved by introducing the statistical features. By comparing the classification performances of SVM, LR, RF, Linear Regression, we can find that RF performs best on the PCA dataset.